Role policy (#8)
* Add json files to store IAM role trust policy

Files can be rendered with jinja2 in order to pass user specific
variables and settings

* Update create_role endpoint

Renders the trust policy from the json template using jinja2.
Add Jinja2 class to help render the templates. Needs to be updated to
pass in further context required.

* Add settings and read vars when creating a role

Config file with a settings object configured to read from a local .env
file. These variables are then read when making the call to create an
aws role.

Update to expect an oidc_user_id to be included when making a request
to create a role

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>